Summarize one variable in a dataset, by another categorical variable

Generate distributional and other statistics for a particular continuous variable, categorized by some discrete variables. Wage by gender for example.

ff_summ_bygroup(
  df,
  vars.group,
  var.numeric,
  str.stats.group = "main",
  str.stats.specify = NULL,
  boo.overall.stats = TRUE
)

Arguments

df
dataframe input dataframe of interest

vars.group
list of strings containing grouping variables, could be gender and age groups for example

var.numeric
string variable name of continuous quantitative variable to summarize

str.stats.group
string what type of statistics to consider see line 31 and below

Value

a list of various variables

df_table_grp_stats - A dataframe where each row is a combination of categories, and columns are categories and statistics

df_row_grp_stats - A single row with all statistics

df_overall_stats - A dataframe with non-grouped overall summaries

df_row_stats_all - A named list of all statistics generated

References

https://fanwangecon.github.io/REconTools/reference/ff_summ_bygroup.html https://github.com/FanWangEcon/REconTools/blob/master/R/ff_summ_bygroup.R

Examples

data(mtcars)
df_mtcars <- mtcars
df <- df_mtcars
vars.group <- c('am', 'vs')
var.numeric <- 'mpg'
str.stats.group <- 'all'
ls_summ_by_group <- ff_summ_bygroup(df, vars.group, var.numeric, str.stats.group)
df_table_grp_stats <- ls_summ_by_group$df_table_grp_stats
df_row_grp_stats <- ls_summ_by_group$df_row_grp_stats
df_overall_stats <- ls_summ_by_group$df_overall_stats
df_row_stats_all <- ls_summ_by_group$df_row_stats_all
print(df_table_grp_stats)

#> # A tibble: 4 x 12
#> # Groups:   am [2]
#>      am    vs  mean median    sd   IQR   mad   min   max first  last n.distinct
#>   <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>      <int>
#> 1     0     0  15.0   15.2  2.77  2.57  2.30  10.4  19.2  10.4  19.2         10
#> 2     0     1  20.7   21.4  2.47  3.5   3.26  17.8  24.4  17.8  24.4          7
#> 3     1     0  19.8   20.4  4.01  4.22  3.85  15    26    15    26            5
#> 4     1     1  28.4   30.4  4.76  6.35  4.60  21.4  33.9  21.4  33.9          6

print(df_row_grp_stats)

#> # A tibble: 1 x 40
#>   mpg.am.vs.0.0.f… mpg.am.vs.0.0.I… mpg.am.vs.0.0.l… mpg.am.vs.0.0.m…
#>              <dbl>            <dbl>            <dbl>            <dbl>
#> 1             10.4             2.57             19.2             2.30
#> # … with 36 more variables: mpg.am.vs.0.0.max <dbl>, mpg.am.vs.0.0.mean <dbl>,
#> #   mpg.am.vs.0.0.median <dbl>, mpg.am.vs.0.0.min <dbl>,
#> #   mpg.am.vs.0.0.n.distinct <dbl>, mpg.am.vs.0.0.sd <dbl>,
#> #   mpg.am.vs.0.1.first <dbl>, mpg.am.vs.0.1.IQR <dbl>,
#> #   mpg.am.vs.0.1.last <dbl>, mpg.am.vs.0.1.mad <dbl>, mpg.am.vs.0.1.max <dbl>,
#> #   mpg.am.vs.0.1.mean <dbl>, mpg.am.vs.0.1.median <dbl>,
#> #   mpg.am.vs.0.1.min <dbl>, mpg.am.vs.0.1.n.distinct <dbl>,
#> #   mpg.am.vs.0.1.sd <dbl>, mpg.am.vs.1.0.first <dbl>, mpg.am.vs.1.0.IQR <dbl>,
#> #   mpg.am.vs.1.0.last <dbl>, mpg.am.vs.1.0.mad <dbl>, mpg.am.vs.1.0.max <dbl>,
#> #   mpg.am.vs.1.0.mean <dbl>, mpg.am.vs.1.0.median <dbl>,
#> #   mpg.am.vs.1.0.min <dbl>, mpg.am.vs.1.0.n.distinct <dbl>,
#> #   mpg.am.vs.1.0.sd <dbl>, mpg.am.vs.1.1.first <dbl>, mpg.am.vs.1.1.IQR <dbl>,
#> #   mpg.am.vs.1.1.last <dbl>, mpg.am.vs.1.1.mad <dbl>, mpg.am.vs.1.1.max <dbl>,
#> #   mpg.am.vs.1.1.mean <dbl>, mpg.am.vs.1.1.median <dbl>,
#> #   mpg.am.vs.1.1.min <dbl>, mpg.am.vs.1.1.n.distinct <dbl>,
#> #   mpg.am.vs.1.1.sd <dbl>

print(df_overall_stats)

#>   mpg.mean mpg.median   mpg.sd mpg.IQR mpg.mad mpg.min mpg.max mpg.first
#> 1 20.09062       19.2 6.026948   7.375 5.41149    10.4    33.9        21
#>   mpg.last mpg.n_distinct
#> 1     21.4             25

print(df_row_stats_all)

#> $mpg.am.vs.0.0.first
#> [1] 10.4
#> 
#> $mpg.am.vs.0.0.IQR
#> [1] 2.575
#> 
#> $mpg.am.vs.0.0.last
#> [1] 19.2
#> 
#> $mpg.am.vs.0.0.mad
#> [1] 2.29803
#> 
#> $mpg.am.vs.0.0.max
#> [1] 19.2
#> 
#> $mpg.am.vs.0.0.mean
#> [1] 15.05
#> 
#> $mpg.am.vs.0.0.median
#> [1] 15.2
#> 
#> $mpg.am.vs.0.0.min
#> [1] 10.4
#> 
#> $mpg.am.vs.0.0.n.distinct
#> [1] 10
#> 
#> $mpg.am.vs.0.0.sd
#> [1] 2.774396
#> 
#> $mpg.am.vs.0.1.first
#> [1] 17.8
#> 
#> $mpg.am.vs.0.1.IQR
#> [1] 3.5
#> 
#> $mpg.am.vs.0.1.last
#> [1] 24.4
#> 
#> $mpg.am.vs.0.1.mad
#> [1] 3.26172
#> 
#> $mpg.am.vs.0.1.max
#> [1] 24.4
#> 
#> $mpg.am.vs.0.1.mean
#> [1] 20.74286
#> 
#> $mpg.am.vs.0.1.median
#> [1] 21.4
#> 
#> $mpg.am.vs.0.1.min
#> [1] 17.8
#> 
#> $mpg.am.vs.0.1.n.distinct
#> [1] 7
#> 
#> $mpg.am.vs.0.1.sd
#> [1] 2.471071
#> 
#> $mpg.am.vs.1.0.first
#> [1] 15
#> 
#> $mpg.am.vs.1.0.IQR
#> [1] 4.225
#> 
#> $mpg.am.vs.1.0.last
#> [1] 26
#> 
#> $mpg.am.vs.1.0.mad
#> [1] 3.85476
#> 
#> $mpg.am.vs.1.0.max
#> [1] 26
#> 
#> $mpg.am.vs.1.0.mean
#> [1] 19.75
#> 
#> $mpg.am.vs.1.0.median
#> [1] 20.35
#> 
#> $mpg.am.vs.1.0.min
#> [1] 15
#> 
#> $mpg.am.vs.1.0.n.distinct
#> [1] 5
#> 
#> $mpg.am.vs.1.0.sd
#> [1] 4.008865
#> 
#> $mpg.am.vs.1.1.first
#> [1] 21.4
#> 
#> $mpg.am.vs.1.1.IQR
#> [1] 6.35
#> 
#> $mpg.am.vs.1.1.last
#> [1] 33.9
#> 
#> $mpg.am.vs.1.1.mad
#> [1] 4.59606
#> 
#> $mpg.am.vs.1.1.max
#> [1] 33.9
#> 
#> $mpg.am.vs.1.1.mean
#> [1] 28.37143
#> 
#> $mpg.am.vs.1.1.median
#> [1] 30.4
#> 
#> $mpg.am.vs.1.1.min
#> [1] 21.4
#> 
#> $mpg.am.vs.1.1.n.distinct
#> [1] 6
#> 
#> $mpg.am.vs.1.1.sd
#> [1] 4.757701
#> 
#> $mpg.mean
#> [1] 20.09062
#> 
#> $mpg.median
#> [1] 19.2
#> 
#> $mpg.sd
#> [1] 6.026948
#> 
#> $mpg.IQR
#> [1] 7.375
#> 
#> $mpg.mad
#> [1] 5.41149
#> 
#> $mpg.min
#> [1] 10.4
#> 
#> $mpg.max
#> [1] 33.9
#> 
#> $mpg.first
#> [1] 21
#> 
#> $mpg.last
#> [1] 21.4
#> 
#> $mpg.n_distinct
#> [1] 25
#>