This function cleans rmd: creates subfolder with pdf, html and R

knit rmd normally generates pdf and html in the same folder. This deletes those files. The file generates a new subfolder by default called htmlpdfr that stores the knit pdf and html files, along with curled R file.

The RMD files should not clear all at the top. That would lead to deleting the rest of the string paths to be searched over.

ff_sup_clean_rmd(
  ar_spt_root = c("C:/Users/fan/R4Econ/amto", "C:/Users/fan/R4Econ/math"),
  ar_spn_skip = c("matrix", "tibble"),
  st_folder_pdf = "/htmlpdfr/",
  st_folder_html = "/htmlpdfr/",
  st_folder_R = "/htmlpdfr/",
  bl_gen_if_git_old = FALSE,
  bl_recursive = TRUE,
  bl_verbose = TRUE,
  bl_test = FALSE
)

Arguments

ar_spn_skip
array a string array of names, if path found contains any of the string in the array, will skip.

st_folder_pdf
string subfolder where to store the pdf file, if just empty, store in rmd folder

st_folder_html
string subfolder where to store the html file, if just empty, store in rmd folder

st_folder_R
string subfolder where to store the R file, if just empty, store in rmd folder

bl_gen_if_git_old
boolean if true then even if RMD files do not have git status change, still update pdf and html files.

bl_recursive
boolean if to search in folders recursively

bl_test
boolean if testing, meaning do not generate pdf html, just see which files are been considered included, searched and found

ls_spt_root
array a string array of path roots in which to search for Rmd files to knit

Value

a list of string paths of files generated

ls_spt_pdf_generated - a list of pdf file names

ls_spt_html_generated - a list of html file names

ls_spt_R_generated - a list of R file names generated

#'

References

https://fanwangecon.github.io/REconTools/reference/ff_sup_clean_rmd.html https://github.com/FanWangEcon/REconTools/blob/master/R/ff_sup_inout.R

Examples

ff_sup_clean_rmd()

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/amto/_main.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/amto/_main.Rmd

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/amto/array/_main.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/amto/array/_main.Rmd

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/amto/array/fs_ary_basics.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/amto/array/fs_ary_basics.Rmd

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/amto/array/fs_ary_generate.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/amto/array/fs_ary_generate.Rmd

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/amto/array/fs_ary_mesh.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/amto/array/fs_ary_mesh.Rmd

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/amto/array/fs_ary_string.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/amto/array/fs_ary_string.Rmd

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/amto/list/_main.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/amto/list/_main.Rmd

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/amto/list/fs_lst_basics.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/amto/list/fs_lst_basics.Rmd

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/math/_main.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/math/_main.Rmd

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/math/combinatorics/_main.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/math/combinatorics/_main.Rmd

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/math/differentiation/_main.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/math/differentiation/_main.Rmd

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/math/func_ineq/_main.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/math/func_ineq/_main.Rmd

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/math/func_ineq/fs_atkinson_ces.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/math/func_ineq/fs_atkinson_ces.Rmd

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/math/func_ineq/fs_gini_disc.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/math/func_ineq/fs_gini_disc.Rmd

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/math/func_ineq/fs_youden.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/math/func_ineq/fs_youden.Rmd

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/math/integration/_main.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/math/integration/_main.Rmd

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/math/integration/fs_integrate_normal.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/math/integration/fs_integrate_normal.Rmd

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/math/solutions/_main.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/math/solutions/_main.Rmd

#> Warning: running command 'git status -s C:/Users/fan/R4Econ/math/solutions/fs_solu_x_lin.Rmd' had status 128

#> NO CHNGE: C:/Users/fan/R4Econ/math/solutions/fs_solu_x_lin.Rmd

#> Generated pdf files:

#> [1] ""
#>    
#> "" 

#> 

#> Generated html files:

#> [1] ""
#>    
#> "" 

#> 

#> Generated R files:

#> [1] ""
#>    
#> "" 

#> 

#> $ls_spt_pdf_generated
#> character(0)
#> 
#> $ls_spt_html_generated
#> character(0)
#> 
#> $ls_spt_R_generated
#> character(0)
#>