Modify R4Econ (and others) Rmd files, add Numbered Sections for HTML, and higher Hierarchy

Modify R4Econ Rmd files, add Numbered Sections for HTML, and higher Hierarchy. Was not able to, for HTML file, use pandoc arguments to add in TOC. And Need to have two hierarchies for RMD's individual specific pdf and html files, and the joint bookdown file combining all.

ff_sup_clean_rmd_mod(st_file, it_hierachy_shift = 2, it_toc_depth = 3)

Arguments

st_file

string texts contents from a R4Econ RMD file.

it_hierachy_shift

rmd modification level of promotion from bookdown hierachy to own file hierarchy

it_toc_depth

rmd own file outputs toc levels to show

Value

modified texts

Author

Fan Wang, http://fanwangecon.github.io