Cumulative sum value from each group from the latest date of availability

svr_id are individuals, observed at multiple points; svr_x is date of observation; svr_y is some value observed at each time point. Imagine these are SAT scores for individuals. we want to calculate the sum/avg etc cumulatively at each date for all individuals, but only using their last score.

ff_panel_cumsum_grouplast(
  df,
  svr_id,
  svr_x,
  svr_y,
  svr_cumsumtop = "y_cumsum",
  stat = "sum"
)

Arguments

df
dataframe

svr_id
string name of individual id variable

svr_x
string name of the ranking variable

svr_y
string name of the value variable

svr_cumsumtop
string variable name that stores the cumulative summed variable

Value

a dataframe with the cumulative summed/averaged etc column up to each row

References

https://fanwangecon.github.io/REconTools/reference/ff_panel_cumsum_grouplast.html https://fanwangecon.github.io/REconTools/articles/fv_panel_cumsum_grouplast.html https://github.com/FanWangEcon/REconTools/blob/master/R/ff_panel_cumsum.R

Examples

library(tidyr)
library(dplyr)
library(tibble)
it_N <- 5
ar_it_M <- sample(1:10, it_N, replace = TRUE)
ar_it_M_ID <- sample(1:it_N, it_N, replace = FALSE)
tb_combine <- as_tibble(cbind(ar_it_M, ar_it_M_ID)) %>% rowid_to_column(var = "id")
tb_long <- tb_combine %>% uncount(ar_it_M)
tb_long <- tb_long %>% add_column(xrand = runif(dim(tb_long)[1])) %>% arrange(xrand) %>% mutate(x = row_number())
tb_long <- tb_long %>% arrange(id, x) %>% group_by(id) %>% mutate(rank_l = row_number())
df <- tb_long %>% select(id, x) %>% add_column(y = runif(dim(tb_long)[1])) %>% arrange(id,x) %>% group_by(id) %>% mutate(y = cumsum(y))
print(df)

#> # A tibble: 20 x 3
#> # Groups:   id [5]
#>       id     x     y
#>    <int> <int> <dbl>
#>  1     1     2 0.411
#>  2     2     9 0.163
#>  3     3     1 0.614
#>  4     3     5 0.965
#>  5     3    11 1.57 
#>  6     3    13 2.06 
#>  7     3    14 2.56 
#>  8     4     3 0.804
#>  9     4    15 1.23 
#> 10     4    17 1.53 
#> 11     5     4 0.951
#> 12     5     6 1.63 
#> 13     5     7 2.30 
#> 14     5     8 2.33 
#> 15     5    10 3.13 
#> 16     5    12 3.28 
#> 17     5    16 4.12 
#> 18     5    18 5.12 
#> 19     5    19 5.98 
#> 20     5    20 6.44 

ff_panel_cumsum_grouplast(df, svr_id='id', svr_x='x', svr_y='y', svr_cumsumtop = 'y_cumsum', stat='sum')

#> # A tibble: 20 x 4
#> # Groups:   id [5]
#>       id     x     y y_cumsum
#>    <int> <int> <dbl>    <dbl>
#>  1     3     1 0.614    0.614
#>  2     1     2 0.411    1.03 
#>  3     4     3 0.804    1.83 
#>  4     5     4 0.951    2.78 
#>  5     3     5 0.965    3.13 
#>  6     5     6 1.63     3.81 
#>  7     5     7 2.30     4.48 
#>  8     5     8 2.33     4.51 
#>  9     2     9 0.163    4.67 
#> 10     5    10 3.13     5.47 
#> 11     3    11 1.57     6.07 
#> 12     5    12 3.28     6.23 
#> 13     3    13 2.06     6.72 
#> 14     3    14 2.56     7.22 
#> 15     4    15 1.23     7.65 
#> 16     5    16 4.12     8.49 
#> 17     4    17 1.53     8.79 
#> 18     5    18 5.12     9.79 
#> 19     5    19 5.98    10.6  
#> 20     5    20 6.44    11.1